requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
djangorestframework==3.15.1
fonttools==4.53.0
gunicorn==22.0.0
idna==3.7
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.1
wheel==0.43.0
